% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumreturns.R
\name{cumreturns}
\alias{cumreturns}
\title{Calculate cumulative returns to an asset or cumulative value of investment in an asset.}
\usage{
cumreturns(returns, growth = TRUE, principal = 1L)
}
\arguments{
\item{returns}{Numerical vector of daily returns to an asset.}

\item{growth}{Logical. TRUE: calculate cumulative growth rate. FALSE: calculate cumulative value of investment. Defaults to TRUE.}

\item{principal}{Numerical. Value of the principal investment. Only used if `growth = FALSE`.}
}
\value{
Numerical vector.
}
\description{
Calculate cumulative returns to an asset or cumulative value of investment in an asset.
}
