% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stonk.R
\name{get_stonk}
\alias{get_stonk}
\title{Fetch asset data and calculate cumulative returns. Mostly a wrapper for `tidyquant::tq_get()`.}
\usage{
get_stonk(
  ticker,
  from = "2007-01-01",
  to = Sys.Date(),
  returns = TRUE,
  loud = TRUE,
  ...
)
}
\arguments{
\item{ticker}{Character. An asset (e.g., "AAPL"). Call multiple assets by putting them in a vector.}

\item{from}{Character. Start date. Defaults to "2007-01-01".}

\item{to}{Character. End date. Defaults to today.}

\item{returns}{Logical. Calculate cumulative returns?}

\item{loud}{Logical. Loud!}

\item{...}{additional arguments for`tidyquant::tq_get()`.}
}
\value{
tibble of asset data.
}
\description{
Fetch asset data and calculate cumulative returns. Mostly a wrapper for `tidyquant::tq_get()`.
}
