% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim.R
\name{plot_sim}
\alias{plot_sim}
\title{Plot distribution of simulated portfolio returns.}
\usage{
plot_sim(.data, normalized = FALSE, cumulative = FALSE, ...)
}
\arguments{
\item{.data}{a data frame from `simulate_returns()`.}

\item{normalized}{Logical. Plot a normalized histogram?}

\item{cumulative}{Logical. Overlay the cumulative distribution?}

\item{...}{additional ggplot arguments.}
}
\value{
a ggplot object.
}
\description{
Plot distribution of simulated portfolio returns.
}
