% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_port.R
\name{plot_sim_port}
\alias{plot_sim_port}
\title{Plot distribution of simulated portfolio returns.}
\usage{
plot_sim_port(
  .data,
  which_portfolio,
  adjust_start = 0,
  adjust_end = 0,
  prices = FALSE,
  ...
)
}
\arguments{
\item{.data}{a data frame from `simulate_returns()`.}

\item{which_portfolio}{Integer or character. Choose a portfolio id (integer) or choose "max" or "min".}

\item{adjust_start}{Integer. Adjust the start date when pulling asset data. Defaults to zero.}

\item{adjust_end}{Integer. Adjust the end date when pulling asset data. Defaults to zero.}

\item{prices}{Logical. Plot asset prices instead of cumulative returns?}

\item{...}{additional arguments to `get_stonk()` (which get passed on to `tidyquant::tq_get()`.}
}
\value{
a ggplot object.
}
\description{
Plot distribution of simulated portfolio returns.
}
