% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_returns.R
\name{simulate_returns}
\alias{simulate_returns}
\title{Simulate portfolios.}
\usage{
simulate_returns(
  .data,
  n_days_in_market = 365,
  cost = 0,
  n_iterations = 10000,
  loud = TRUE
)
}
\arguments{
\item{.data}{a data frame of an asset.}

\item{n_days_in_market}{Integer. How long is the asset investment.}

\item{cost}{Numeric. Cost or expense ratio to asset.}

\item{loud}{Logical. Loud!}

\item{n_iterations.}{Integer. Number of iterations to run the simulation.}
}
\value{
data frame of simulation results.
}
\description{
Simulate portfolios.
}
