% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sequence.R
\name{slice_sequence}
\alias{slice_sequence}
\title{Slice a random, *ordered* sequence from the data. Similar in spirit to `dplyr::slice_sample()`.}
\usage{
slice_sequence(.data, n)
}
\arguments{
\item{.data}{a data frame.}

\item{n}{Integer. Sample size.}
}
\value{
a sampled data frame.
}
\description{
Slice a random, *ordered* sequence from the data. Similar in spirit to `dplyr::slice_sample()`.
}
