% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_stonk.R
\name{theme_stonk}
\alias{theme_stonk}
\title{A ggplot theme.}
\usage{
theme_stonk(base_size = 16, family = "sans", face = "bold", hjust = 0)
}
\arguments{
\item{base_size}{Integer. Base size of font.}

\item{family}{Character. Font family. Defaults to "sans".}

\item{face}{Character. Font face. Defaults to "bold".}

\item{hjust}{Integer. Adjustment of y axis title position. Defaults to 0.}
}
\value{
ggplot theme.
}
\description{
A ggplot theme.
}
